# Fix performance issue reimport file reload scene