… editor/editor_node.cpp