- Fixes #95104

This PR should significantly improve performance when reimporting assets with the auto-reload of scenes.

### What has been done:
- **Addition and removal of the reloaded scene in `EditorNode::preload_reimporting_with_path_in_edited_scenes`**: It takes 7-8 seconds on my PC just to remove and re-add the scene. This part should not require the scene to be in the tree to load the modified properties and nodes to reimport. I had to modify `EditorNode::get_preload_scene_modification_table` to use `get_transform` instead of `get_relative_transform`. I don't know why `get_relative_transform` was used, but in my tests, it works fine with `get_transform`. Additionally, when the transform is restored, it uses `set_transform`.
- **Adjustment in `EditorNode::reload_instances_with_path_in_edited_scenes`**: The addition and removal of the reloaded scene have been minimized. The current edited scene should almost never be removed unless another scene that needs to be reloaded has the same name.
- **Modification in `EditorNode::_set_current_scene_nocheck`**: The addition and removal of the reloaded scene have been removed if the same scene is reloaded.
- **Replacement of `EditorNode::update_node_reference_modification_table_for_node`**: This method has been replaced by `EditorNode::get_preload_modifications_reference_to_nodes`, which now runs in the preload and loads only modified properties that reference reimported nodes. Previously, it loaded every property modification in the scene, taking about 10 seconds. Now, it takes around 7 seconds in debug mode. The main issue was the call to `p_node->get_property_list` for each node in the scene. I have not found a better solution to optimize this.
- **Added progress popups**: These display visual feedback to the user before and during the reloading process.

### Tested
- I retested with the MRP (https://github.com/Jamsers/Crater-Province-Level/tree/test-4.2).
- I retested with the test project from #90531 with some minor modifications: [test-godot-blender-reimport-missing-node-v5.zip](https://github.com/user-attachments/files/16517581/test-godot-blender-reimport-missing-node-v5.zip).

### Known issues
- On my PC, opening Godot on startup with the MRP is very slow. Commenting out the method `Node3DEditor::_request_gizmo` removes 15 seconds from the startup time.
- I created issue #95221 for signals that never disconnect and accumulate on nodes.
- Adding and removing scenes with many nodes is slow. Just switching scene tabs can take 5-6 seconds. Most of the time is spent in `add_child` and `remove_child`.

I will not work on these issues for now, as I think the current performance of this PR is acceptable.

### Time
Reloading an asset in the MRP (debug dev build):
- Before: 5 min. +
- Now: 10-12 sec.

Reloading an asset in the MRP (prod build):
- Before (4.3 rc2): 3 min. +
- Now (artifact from the PR): 4-6 sec.

With the artifact from the PR:

https://github.com/user-attachments/assets/9ac6be44-3607-4269-b8b2-67ef4942ff1d