@Calinou 2024-08-07
> On my PC, opening Godot on startup with the MRP is very slow. Commenting out the method Node3DEditor::_request_gizmo removes 15 seconds from the startup time.

See https://github.com/godotengine/godot/issues/94648.