@akien-mga 2024-08-07
Thanks so much for working on this @Hilderin, you're once again saving the 4.3 release :1st_place_medal: 
This regression is the main release blocker left IMO, so once we have this merged we should hopefully be good to go :crossed_fingers: 

I'll try to give this a quick test and superficial review, and aim to merge it today so I can release `4.3.rc3` tomorrow as hopefully the "final FINAL v3 final.psd" RC!