@akien-mga 2024-08-07
- Test with https://github.com/Nodragem/twin-stick-shooter-starter-kit which I had laying around (I'm on 5G now so can't download big MRPs).

Modifying all `.gltf` files to add a space at the end and thus trigger a reimport (`for file in $(find -name "*.gltf"); do echo " " >> $file; done`):

Current `master`: 13 s 870 ms (dev build), 5 s 380 ms (optimized build)
This PR: 5 s 410 ms (dev build), 3 s 870 ms (optimized build)

- Test with (private, for now) W4 demo Project Genesis, modifying the massive `FortifiedCastleTown.gltf` file to add an empty space and trigger a reimport.

Current `master`: More than 8 min, I gave up (optimized build)
This PR: 1 min 34 s (optimized build)

Seems pretty drastic!

I've also done a cursory review of the code, looks fine to me overall, but I'm no expert in this part of the engine.