@clayjohn 2024-08-07
I have tested this PR and confirm that it fixes the performance and brings things in line with expectations (3+ minutes down to 7 seconds). 

On my device, initial import of the entire project is about 45 seconds (25 seconds to import + 20 seconds to open the scene). 

On master re-importing a single mesh took 3+ minutes and with this PR the one i tested took 7 seconds.